Fix style of overview component
The event overview component had its search and tab bar flying around at
the top. This is now closer to the material design specification:

- https://material.io/guidelines/layout/principles.html
- https://material.io/guidelines/components/tabs.html

There remains an issue with the box shadow of the toolbar overlapping this
sub-navigation.

Also added a variety of doc comments and replaced the computed
group-by-type value with a quasi-cached one (maintaining UAP).