Update question-tool style
The previous design did not have a good fit with the overall material
design of the app. The updated design follows the material design spec
more closely.

The author is now set to the users username on creation instead of you to
prevent weired occurences where some of the users comments had her name,
while some had 'You'.
Author-names are now also links to the authors page.

Date-formatting is now such that everything under a minute is 'Just now',
everything else is the formatted date-time.