Update style of event-file-upload
Previously upload-tool, the component has been renamed to
event-file-upload. It requires an event and cannot be used with any other
sort of data structure, so to have that reflected in its names it was renamed.

There were some changes in the design to make it closer to the material
design.
The order of the elements has been changed to correspond to the logical
order of the steps taken: First choose a file, then review its meta-info,
then submit (compared to previously having the meta-info above the file
selector).

The warning/info box has been styled to clearly express its warning
nature.

A <label> tab was introduced to increase the interactive area of the
element.