Refactor / style / document frontend components (#141)
* Update models

- Make their implementation more consistent
- Remove recursive merging in favor of explicit constructor args
- Add doc comments for the models

* Add doc to nginx config file

* Update frontend config

* Update import order in app files

* Remove unused app.resolver

* Update app component

* Fix minor issue with form validators test case

* Update app manifest

* Restructure global stylesheets

The global stylesheets had become a bit unorganized with layout- and
component specific styles in the main.scss. This is amended by moving
rules to their respective files.
Additionally some unnecessary classes are removed.

* Propagate conference.model change

* Propagate div model change

* Fix import order; remove event-filter.pipe

* Add doc comments; remove unused services

* Cleanup and document auth components

* Replace grouping filter with single-traversal-getter

The event-overview had a Pipe for filtering events by their type.
This had several disadvantages: One needed to know all types in advance,
the events had to be traversed multiple times etc.

This is replaced by a single-traversal that collects all existing types
and groups the events accordingly.

* Fix auth API changes in backend

* Clean up navigation components

- Move style from navigation-user.component.ts to global _navigation.scss.
  This was done to not have CSS mixen in with the other code.
- Replace button of NaivgationComponent with a simple span, since it only
  displays data and is not interactive.
- Remove some unused code.

* Fix missing of in event overview template

A missing of in a let-of loop was causing no events to be displayed.
The lack of null-catching in Person.fromAPI also was an issue and is fixed
by this commit.

* Fix VLDB2016 demo data insertion script

The script contained the incorrect column name "roomname" instead of
"roomid"

* Add search to voting list

Due to the length of the list, voting on specific papers may become
cumbersome. This can be alleviated by adding a search/filter option.

Also added some doc comments and updated the CSS to use flex-box instead
of floats.

* Add AuthGuard to my-schedule

Make personal schedule no longer accessible when not logged in.

* Fix layout; doc comments in my-schedule

* Fix style of overview component

The event overview component had its search and tab bar flying around at
the top. This is now closer to the material design specification:

- https://material.io/guidelines/layout/principles.html
- https://material.io/guidelines/components/tabs.html

There remains an issue with the box shadow of the toolbar overlapping this
sub-navigation.

Also added a variety of doc comments and replaced the computed
group-by-type value with a quasi-cached one (maintaining UAP).

* Fix style issues in event-list.component

Code-style-wise most notably the camelCase class names.
The layouting of the card content also used some weired margins and
paddings to achieve the desired layout.
This has been replaced with mostly flex-box layout.
Some of the HTML tags were also replaced with semantically more
appropriate alternatives.

* Update event-view style

- Replaced some markup with semantically more appropriate alternatives
  (e.g. tabular data in tables, removed md-list when content was not a
  list etc.)
- Update stylesheets to preserver original design. There were some
  changes, mostly replacing akward layouting with flex-box and some
  signifiers for links
- Add `formatMoment` pipe to format momentjs time objects

* Style person-view.component

The person-view component had a few unstyled elements. They have been
given styling. In addition HTML semantics have been improved.

* Update question-tool style

The previous design did not have a good fit with the overall material
design of the app. The updated design follows the material design spec
more closely.

The author is now set to the users username on creation instead of you to
prevent weired occurences where some of the users comments had her name,
while some had 'You'.
Author-names are now also links to the authors page.

Date-formatting is now such that everything under a minute is 'Just now',
everything else is the formatted date-time.

* Update style of event-file-upload

Previously upload-tool, the component has been renamed to
event-file-upload. It requires an event and cannot be used with any other
sort of data structure, so to have that reflected in its names it was renamed.

There were some changes in the design to make it closer to the material
design.
The order of the elements has been changed to correspond to the logical
order of the steps taken: First choose a file, then review its meta-info,
then submit (compared to previously having the meta-info above the file
selector).

The warning/info box has been styled to clearly express its warning
nature.

A <label> tab was introduced to increase the interactive area of the
element.